Request for changes corrected
