+++ python/setup_te.py
@@ -61,5 +61,5 @@ setup(
         'Programming Language :: Python :: 3',
         'Operating System :: POSIX :: Linux',
     ],
-    python_requires='>=3.6',
+    python_requires='>=3.11',
 )
